Change type of PlusAddressSettingSpecifics::int_value
There's no reason to restrict it to int32. This CL changes the type of
the field to int64 instead. Since the type hasn't launched yet, this is
safe.

Bug: b/342089839
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5578580
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1307351}

+++ components/plus_addresses/settings/plus_address_setting_sync_bridge_unittest.cc
@@ -29,7 +29,7 @@ namespace {
 
 syncer::EntityData CreateSettingEntity(
     const std::string& name,
-    absl::variant<bool, std::string, int32_t> value) {
+    absl::variant<bool, std::string, int64_t> value) {
   syncer::EntityData entity;
   sync_pb::PlusAddressSettingSpecifics* specifics =
       entity.specifics.mutable_plus_address_setting();
@@ -38,7 +38,7 @@ syncer::EntityData CreateSettingEntity(
       base::Overloaded{
           [&](bool value) { specifics->set_bool_value(value); },
           [&](const std::string& value) { specifics->set_string_value(value); },
-          [&](int32_t value) { specifics->set_int_value(value); }},
+          [&](int64_t value) { specifics->set_int_value(value); }},
       value);
   return entity;
 }
